Midlertidig implementering av arbeidsgiverNorge
