+++ src/ducks/saksopplysninger.js
@@ -55,6 +55,11 @@ export const ArbeidsforholdSelector = createSelector(
   arbeidsforhold => arbeidsforhold
 );
 
+export const ArbeidsgiverSelector = createSelector(
+  state => state.saksopplysninger.data.arbeidsforhold,
+  arbeidsforhold => arbeidsforhold
+);
+
 export const ArbeidsforholdetSelector = createSelector(
   (state, arbeidsforholdID) => arbeidsforholdID,
   state => state.saksopplysninger.data.arbeidsforhold || [],
