+++ src/felles-komponenter/arbeidsgiverNorge.less
@@ -1,12 +1,8 @@
 @import '../utils/mixins';
 
 .arbeidsgiverNorge {
-
-  .person__seksjon {
-    .seksjonmixin();
-  }
-
-  .person__detaljer {
+  .arbeidsgiver__detaljer {
     .beskrivliste-mixin('col');
   }
+
 }
